+++ src/App.js
@@ -13,9 +13,7 @@ import Oops from './components/Error/Oops';
 import NotFoundPage from './components/Error/NotFoundPage';
 import FileListContainer from "./components/Repository/FileListContainer";
 import packagejson from '../package.json';
-import AppSearchAPIConnector from "@elastic/search-ui-app-search-connector";
-import { SearchProvider } from "@elastic/react-search-ui";
-import ReportCard from './components/ReportCard/ReportCard';
+import ReportCardContainer from './components/ReportCard/ReportCardContainer';
 
 const cacheStore = window.sessionStorage.getItem('hyrda-redux-store');
 const initialState = cacheStore ? JSON.parse(cacheStore) : loadedState;
@@ -50,71 +48,6 @@ store.subscribe(function () {
 
 store.subscribe(saveState);
 
-const connector = new AppSearchAPIConnector({
-  searchKey: process.env.REACT_APP_SEARCH_KEY,
-  endpointBase: process.env.REACT_APP_SEARCH_ENDPOINT,
-  engineName: "atlas-repository",
-  cacheResponses: false
-})
-
-const searchConfig = {
-  apiConnector: connector,
-  searchQuery: {
-      disjunctiveFacets: [
-        "id",
-        "data_format",
-        "access",
-        "redcap_id",
-        "file_name",
-        "data_category",
-        "workflow_type",
-        "package_id",
-        "platform",
-        "file_size",
-        "file_id",
-        "data_type",
-        "dois",
-        "experimental_strategy",
-        "sex",
-        "age_binned",
-        "tissue_type",
-        "sample_type",
-        "tissue_source",
-        "protocol",
-        "release_version"
-      ],
-      facets: {
-        id: { type: "value", size: 500},
-        data_format: { type: "value", size: 250},
-        access: { type: "value", size: 250},
-        redcap_id: { type: "value", size: 500},
-        file_name: { type: "value", size: 250},
-        data_category: { type: "value", size: 250},
-        workflow_type: { type: "value", size: 250},
-        package_id: { type: "value", size: 250},
-        platform: { type: "value", size: 250},
-        file_size: { type: "value", size: 250},
-        file_id: { type: "value", size: 250},
-        data_type: { type: "value", size: 250},
-        dois: { type: "value", size: 250},
-        experimental_strategy: { type: "value", size: 250},
-        sex: { type: "value", size: 250},
-        age_binned: { type: "value", size: 250},
-        tissue_type: { type: "value", size:250},
-        sample_type: { type: "value", size: 250},
-        tissue_source: { type: "value", size: 250},
-        protocol: { type: "value", size: 250},
-        release_version: { type: "value", size: 250}
-      }
-  },
-  initialState: {
-    resultsPerPage: 20,
-    current: 1
-  },
-  trackUrlState: true,
-  alwaysSearchOnInitialLoad: true
-}
-
 class App extends Component {
   componentWillMount() {
     logPageView(window.location, '');
@@ -123,20 +56,18 @@ class App extends Component {
   render() {
     return (
       <Provider store={store}>
-        <SearchProvider config={searchConfig}>
           <BrowserRouter history={history} basename={packagejson.baseURL}>
             <ErrorBoundaryContainer>
               <NavBar app='atlas' />
               <Switch>
                 <Route exact path="/" component={FileListContainer} store={store} />
-                <Route exact path="/report" component={ReportCard} store={store}/>
+                <Route exact path="/report" component={ReportCardContainer} store={store}/>
                 <Route exact path="/oops" component={Oops} />
                 <Route path='*' component={NotFoundPage} />
               </Switch>
               <NavFooter app='atlas' />
             </ErrorBoundaryContainer>
           </BrowserRouter>
-        </SearchProvider>
       </Provider>
     );
   }